# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kazaa API",
        "description": "API documentation for Kazaa system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/{locale}/user/branches/{branch_id}": {
            "get": {
                "tags": [
                    "Branch Data"
                ],
                "summary": "Get Main Data Of Branch",
                "description": "Return 'Vendor Data' ,'Branch Data' ",
                "operationId": "333d3ed46831486cc42b1c222e1b0af4",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Language code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    },
                    {
                        "name": "branch_id",
                        "in": "path",
                        "description": "the id of branch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Dokki branch"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "Dokki branch"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "data": {
                                        "id": 1,
                                        "name": "Dokki branch",
                                        "slug": "Dokki branch",
                                        "current_Status_opening_hours": "open",
                                        "phone_number": "1233665566666",
                                        "address": "نعلﻻلرتﻻتﻻ",
                                        "google_place_link": "https://maps.app.goo.gl/KaS8Czd6Qmy8suus8",
                                        "vendor_data": {
                                            "id": 1,
                                            "brand_name": "shams",
                                            "logo": "image",
                                            "banar": "image",
                                            "slogan": "Dishes that please the palate"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "AppAPIKey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            },
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}